// src/enums.ts
export enum ListStatus {
    IDLE = "idle",
    LOADING = "loading",
    FAILED = "failed",
}

export enum AddStatus {
    IDLE = "idle",
    ADDING = "adding",
    SUCCESS = "success",
    FAILED = "failed",
}

export enum DeleteStatus {
    IDLE = "idle",
    DELETING = "deleting",
    SUCCESS = "success",
    FAILED = "failed",
}